Directional pad with dynamic components
For games link Mario Party the user can change the
layout for a better playability of the game. Some
games request the user to use just A and B, for
example, so the layout changes to just show the
two buttons.

There are 3 master layouts:

* Two touch buttons
* Three touch buttons
* One directional and 2 touch buttons

With proper keyset changes these 3 sets can cover
all minigames of the game.

Fixed #29.